pyproject.toml
[project]
name = "gherkbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "schlich", email = "ty.schlich@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "gherkin-official>=32.1.2",
    "pydantic>=2.11.5",
]

[project.scripts]
gherkbot = "gherkbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "basedpyright",
    "pyrefly",
    "ty",
    "hypothesis>=6.135.2",
    "pytest-watcher>=0.4.3",
    "ruff>=0.11.13",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib"
]
